.opencode/skills/ppt-audit-polish/pyproject.toml
[project]
name = "ppt-audit-polish"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=11.0.0",
  "PyMuPDF>=1.25.0",
  "python-pptx>=1.0.2",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "scripts"]
testpaths = ["tests"]
